Convert header type to comply with header model

+++ miniscope_io/stream_daq.py
@@ -163,7 +163,9 @@ class StreamDaq:
             reverse_payload_bytes=self.config.reverse_payload_bytes,
         )
 
-        header_data = StreamBufferHeader.from_format(header, self.header_fmt, construct=True)
+        header_data = StreamBufferHeader.from_format(
+            header.astype(int), self.header_fmt, construct=True
+        )
         header_data.adc_scaling = self.config.adc_scale
 
         return header_data, payload
